Update UI to be applicable to bridge.net handling

+++ GeonBit.UI/Source/Animators/TextWaveAnimator.cs
@@ -106,7 +106,7 @@ namespace GeonBit.UI.Animators
         /// </summary>
         private void PerCharacterManipulationFunc(Entities.RichParagraph paragraph, char currChar, int index, ref Color fillColor, ref Color outlineColor, ref int outlineWidth, ref Vector2 offset, ref float scale)
         {
-            float distance = System.Math.Abs((float)index - _currPosition);
+            float distance = (float)System.Math.Abs((float)index - _currPosition);
             offset.Y = -WaveHeight / (1f + ((distance * distance) / WaveLengthFactor));
         }
     }
